Rearranged botch initial hierarchy.

+++ include/arena.h
@@ -1,4 +1,3 @@
-/* $Id: H:/drh/idioms/book/RCS/arena.doc,v 1.10 1997/02/21 19:45:19 drh Exp $ */
 #ifndef ARENA_INCLUDED
 #define ARENA_INCLUDED
 #include "except.h"
@@ -8,10 +7,11 @@ extern const Except_T Arena_NewFailed;
 extern const Except_T Arena_Failed;
 extern T    Arena_new    (void);
 extern void Arena_dispose(T *ap);
-extern void *Arena_alloc (T arena, long nbytes,
+extern void *Arena_alloc (T arena, int nbytes,
 	const char *file, int line);
-extern void *Arena_calloc(T arena, long count,
-	long nbytes, const char *file, int line);
+extern void *Arena_calloc(T arena, int count, int nbytes,
+	const char *file, int line);
 extern void  Arena_free  (T arena);
 #undef T
 #endif
+/* $RCSfile: RCS/arena.doc,v $ $Revision: 1.6 $ */
